[WHD-95] Design: Change club name and account validation page
동아리 이름 검증 및 계좌 인증 화면 UI 변경
- 동아리 설명을 이름과 함께 입력하게 함
- 하단 버튼이 아닌 검증 및 인증 버튼을 따로 만듦
